// app/lib/types.ts
export interface WeatherApiSearchItem {
    id: number;
    name: string;
    region: string;
    country: string;
    lat: number;
    lon: number;
    url: string;
}
// ...
export interface SearchResultItem {
    id: number;
    name: string;
    region: string;
    country: string;
    lat: number;
    lon: number;
}